Refactor blog page components to use BlogPageCard for rendering posts, implement mobile and desktop navigation components, and update footer styles for consistency
