+++ todo
@@ -1,16 +1,7 @@
 
-// TODO: Update happens when you click on a post or project.
-//  As the user types, updates are saved automatically to the device.
-//  There should be a button to preview the work.
-//  The preview is simply redirecting them to client side as usual.
-//  A delete button should be there as well.
-//  Maybe a publish/unpublish button as well.
-//  Don't forget the TinyMCE text editor.
-//  There should also be a settings page to update the picture and social media links that show on the client side.
-//  Also don't forget to implement search.
-// always cache posts or projects being created incase the token expires and they redirected to the login page.
+There should also be a settings page to update the picture and social media links that show on the client side.
 
-Fit admin post title and date together. Everything else should be fine
+Always cache posts or projects being created incase the token expires and they redirected to the login page.
 
 Update the create post/project modal. The title and close button should be well aligned
 
@@ -18,4 +9,4 @@ Also optimize seo
 
 Clean up code later. Most pages need components. Start with admin posts
 
-Investigate toast
+Make admin and client side share the navbar, difference only in content
